backend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
pi==0.1.2
pytz==2020.4
sqlparse==0.4.1
gunicorn
